requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
logzero==1.7.0
numpy==1.25.2
pandas==2.0.3
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.31.0
six==1.17.0
smartapi-python==1.5.4
tabulate==0.9.0
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.6.4
plotly